carapi/requirements.txt
Django==1.11.4
django-filter==1.0.4
djangorestframework==3.6.3
drfdocs==0.0.11
psycopg2==2.7.3
pytz==2017.2
